Move GetClientObjects to src\util
