+++ src/userList2.js
@@ -1,4 +1,5 @@
 import React from "react";
+import { GetClientObjects } from "./util/userDataListUtil";
 
 const UserList2 = ({ userDataList }) => {
   let clientObjectList = GetClientObjects(userDataList);
@@ -33,37 +34,5 @@ const UserList2Vendor = ({ name: vendorName, users }) => (
 );
 
 const UserList2User = ({ name: userName }) => <p>{userName}</p>;
-
-function GetClientObjects(userDataList) {
-  let clientObjectList = [];
-
-  userDataList.forEach(
-    ({ name: userName, client: userClient, vendor: userVendor }) => {
-      // find or create client object
-      let clientObject = clientObjectList.find(
-        ({ name: clientName }) => clientName === userClient
-      );
-      if (clientObject == null) {
-        clientObject = { name: userClient, vendors: [] };
-        clientObjectList.push(clientObject);
-      }
-
-      // find or create vendor object
-      let vendorObject = clientObject.vendors.find(
-        ({ name: vendorName }) => vendorName === userVendor
-      );
-      if (vendorObject == null) {
-        vendorObject = { name: userVendor, users: [] };
-        clientObject.vendors.push(vendorObject);
-      }
-
-      // add user
-      let userObject = { name: userName };
-      vendorObject.users.push(userObject);
-    }
-  );
-
-  return clientObjectList;
-}
 
 export default UserList2;
